[amberscript] split parser test file
The parser test file is getting large and hard to work with. It's
difficult to find what has been tested and where the tests are, along
with causing merge conflicts.

This CL splits the monolithic parser_test file into files based on
what's being parsed.
